Automatically print JSON atoms -- fix #5
There is no reason to push anything but a JSON collection (eg. array or
object) onto the data stack, because the only thing you can do with them
is print them. This commit just automatically prints them instead of
requiring that they are pushed onto the stack and then printed using the
% command.

Compatibility note: The % command continues to be supported but is now
marked as deprecated.