Roll back implicit printing changes -- unfixes #5
The problem with automatic printing is demonstrated in test L46:

    JSON='{"foo":"a","bar":{"x":"b"},"baz":[{"y":"c"},{"y":"d","z":"e"}]}'
    echo "$JSON" |jt baz ^

How many columns should the result have?

We want to guarantee that there will be exactly one column in the output
for each % and ^ command. With the recent changes, the number of columns
depends on the type of the "bar" value as well as the commands given.

With explicit printing via % we can at least guarantee a stable number
of columns of output, with nulls in fields that can't be printed because
of their type.