The ^ command now prints the line index when streaming
Suppose you are using the -s option (eg. one JSON object per line in a
log file). The ^ command now prints the line index (starting from zero)
when called at the root (not in a nested array). The previous behavior
was to exit with an error in this case. When in a nested array the ^
command works as before.